requirements.txt
langchain 
langchain-community
langchain-groq
chromadb
python-dotenv
pypdf
sentence-transformers